+++ demos/react-demo/src/pages/Upload.tsx
@@ -32,7 +32,7 @@ export default function Upload(): JSX.Element {
 
     try {
       // Fetch files using GraphQL query
-      const response = await nhost.graphql.post<GraphqlGetFilesResponse>({
+      const response = await nhost.graphql.request<GraphqlGetFilesResponse>({
         query: `query GetFiles {
             files {
               id